refactor: move cmd/process related methods to new package
