+++ main.go
@@ -3,11 +3,10 @@ package main
 import (
 	"fmt"
 	"log"
+	"maploader/cmd"
 	"maploader/config"
 	"maploader/tar"
 	"maploader/util"
-	"os"
-	"os/exec"
 	"path/filepath"
 	"time"
 
@@ -128,7 +127,7 @@ func changeMap(client mqtt.Client, newMap string) {
 	checkAndHandleErrorWithMqtt(err, client)
 
 	log.Println("stopping processes")
-	stopProcesses()
+	cmd.StopProcesses()
 
 	log.Println("removing current map files")
 	util.RemoveDirContents("/data/ri/")
@@ -147,57 +146,15 @@ func changeMap(client mqtt.Client, newMap string) {
 	}
 
 	log.Println("map change complete, syncing files")
-	excuteCmd("sync")
+	cmd.ExcuteCmd("sync")
 
 	log.Println("restarting processes")
-	startProcesses()
+	cmd.StartProcesses()
 
 	currentMap = newMap
 	publishCurrentMap(client)
 	publishState(client, "idle")
 
-}
-
-func stopProcesses() {
-	excuteCmd("killall", "-9", "valetudo")
-	excuteCmd("sh", "/etc/rc.d/miio.sh", "stop")
-	excuteCmd("killall", "-9", "ava")
-}
-
-func startProcesses() {
-	excuteCmd("sh", "/etc/rc.d/miio.sh")
-	excuteCmd("sh", "/etc/rc.d/ava.sh")
-	startValetudo()
-}
-
-func startValetudo() {
-	devnull, dnerr := os.OpenFile(os.DevNull, os.O_WRONLY, 0755)
-	if dnerr != nil {
-		util.CheckAndHandleError(dnerr)
-	}
-
-	cmd := exec.Command("/data/valetudo")
-	cmd.Stdout = devnull
-	cmd.Env = os.Environ()
-	valetudoConfigEnv := "VALETUDO_CONFIG_PATH=" + config.Getenv("VALETUDO_CONFIG_PATH", "/data/valetudo_config.json")
-	cmd.Env = append(cmd.Env, valetudoConfigEnv)
-	err := cmd.Start()
-
-	if err != nil {
-		util.CheckAndHandleError(err)
-	} else {
-		cmd.Process.Release()
-	}
-}
-
-func excuteCmd(cmdStr string, cmdArgs ...string) {
-
-	cmd := exec.Command(cmdStr, cmdArgs...)
-	err := cmd.Run()
-
-	if err != nil {
-		util.CheckAndHandleError(err)
-	}
 }
 
 func checkAndHandleErrorWithMqtt(err error, client mqtt.Client) {
